Ajout gestion bdd + liste de plats

+++ src/strategy/HappyHour.java
@@ -1,7 +1,7 @@
 package strategy;
 
 public class HappyHour implements TarificationStrategy {
-    public double calculerPrix(double prixInitial) {
-        return prixInitial * 0.8; // 20% de réduction
+    public int calculerPrix(int prixInitial) {
+        return (int) (prixInitial * 0.8); // 20% de réduction
     }
 }
