+++ src/strategy/Reduction.java
@@ -1,7 +1,7 @@
 package strategy;
 
 public class Reduction implements TarificationStrategy {
-    public double calculerPrix(double prixInitial) {
-        return prixInitial * 0.9; // 10% de réduction
+    public int calculerPrix(int prixInitial) {
+        return (int) (prixInitial * 0.9); // 10% de réduction
     }
 }
